+++ __tests__/unit/core-database/__fixtures__/state-storage-stub.ts
@@ -5,6 +5,7 @@ import { Blocks, Interfaces } from "@arkecosystem/crypto";
 export class StateStoreStub implements State.IStateStore {
     public blockchain: any;
     public lastDownloadedBlock: Interfaces.IBlockData | undefined;
+    public genesisBlock: Interfaces.IBlock | undefined;
     public blockPing: any;
     public started: boolean;
     public forkedBlock: Interfaces.IBlock | undefined;
@@ -34,6 +35,14 @@ export class StateStoreStub implements State.IStateStore {
 
     public getLastHeight(): number {
         return 1;
+    }
+
+    public getGenesisBlock(): Interfaces.IBlock {
+        return this.genesisBlock;
+    }
+
+    public setGenesisBlock(block: Interfaces.IBlock): void {
+        this.genesisBlock = block;
     }
 
     public getLastBlock(): Interfaces.IBlock | undefined {