+++ __tests__/unit/core-state/mocks/database.ts
@@ -5,7 +5,6 @@ export const database = {
         findByPublicKey: pubKey => "username",
     },
 
-    commitQueuedQueries: () => undefined,
     buildWallets: () => undefined,
     saveWallets: () => undefined,
     getLastBlock: () => undefined,
@@ -15,7 +14,6 @@ export const database = {
     applyRound: () => undefined,
     getActiveDelegates: () => [],
     restoreCurrentRound: () => undefined,
-    enqueueDeleteBlock: () => undefined,
     getBlocks: () => [],
     getBlock: () => undefined,
     revertBlock: () => undefined,