+++ packages/core-database/src/database-service.ts
@@ -44,6 +44,10 @@ export class DatabaseService implements Database.IDatabaseService {
     }
 
     public async init(): Promise<void> {
+        app.resolvePlugin<State.IStateService>("state")
+            .getStore()
+            .setGenesisBlock(Blocks.BlockFactory.fromJson(Managers.configManager.get("genesisBlock")));
+
         if (process.env.CORE_RESET_DATABASE) {
             await this.reset();
         }
@@ -54,7 +58,11 @@ export class DatabaseService implements Database.IDatabaseService {
 
         Managers.configManager.setHeight(lastBlock.data.height);
 
-        await this.loadBlocksFromCurrentRound();
+        try {
+            await this.loadBlocksFromCurrentRound();
+        } catch (error) {
+            this.logger.warn(`Failed to load blocks from current round: ${error.message}`);
+        }
 
         await this.configureState(lastBlock);
     }
@@ -65,11 +73,14 @@ export class DatabaseService implements Database.IDatabaseService {
     }
 
     public async reset(): Promise<void> {
-        await this.connection.blocksRepository.truncate();
-        await this.connection.roundsRepository.truncate();
-        await this.connection.transactionsRepository.truncate();
+        await this.connection.resetAll();
 
-        await this.saveBlock(Blocks.BlockFactory.fromJson(Managers.configManager.get("genesisBlock")));
+        await this.saveBlock(
+            app
+                .resolvePlugin<State.IStateService>("state")
+                .getStore()
+                .getGenesisBlock(),
+        );
     }
 
     public async applyBlock(block: Interfaces.IBlock): Promise<void> {
@@ -137,24 +148,12 @@ export class DatabaseService implements Database.IDatabaseService {
         await this.connection.buildWallets();
     }
 
-    public async commitQueuedQueries(): Promise<void> {
-        await this.connection.commitQueuedQueries();
-    }
-
-    public async deleteBlock(block: Interfaces.IBlock): Promise<void> {
-        await this.connection.deleteBlock(block);
+    public async deleteBlocks(blocks: Interfaces.IBlockData[]): Promise<void> {
+        await this.connection.deleteBlocks(blocks);
     }
 
     public async deleteRound(round: number): Promise<void> {
         await this.connection.roundsRepository.delete(round);
-    }
-
-    public enqueueDeleteBlock(block: Interfaces.IBlock): void {
-        this.connection.enqueueDeleteBlock(block);
-    }
-
-    public enqueueDeleteRound(height: number): void {
-        this.connection.enqueueDeleteRound(height);
     }
 
     public async getActiveDelegates(
@@ -582,7 +581,12 @@ export class DatabaseService implements Database.IDatabaseService {
         if (!(await this.getLastBlock())) {
             this.logger.warn("No block found in database");
 
-            await this.saveBlock(Blocks.BlockFactory.fromJson(this.config.get("genesisBlock")));
+            await this.saveBlock(
+                app
+                    .resolvePlugin<State.IStateService>("state")
+                    .getStore()
+                    .getGenesisBlock(),
+            );
         }
     }