+++ packages/core-interfaces/src/core-database/database-repository/blocks-repository.ts
@@ -24,7 +24,7 @@ export interface IBlocksRepository extends IRepository {
     }>;
 
     top(count: number): Promise<Interfaces.IBlockData[]>;
-    delete(id: string): Promise<void>;
+    delete(ids: string[], db: any): Promise<void>;
 
     getBlockRewards(): Promise<any>;
 
@@ -32,7 +32,5 @@ export interface IBlocksRepository extends IRepository {
 
     getDelegatesForgedBlocks(): Promise<any>;
 
-    /* TODO: Remove with V1 */
-    findAll(params: ISearchParameters): Promise<IBlocksPaginated>;
     search(params: ISearchParameters): Promise<IBlocksPaginated>;
 }