+++ packages/core-tester-cli/src/commands/command.ts
@@ -72,7 +72,7 @@ export abstract class BaseCommand extends Command {
     protected async make(command): Promise<any> {
         const { args, flags } = this.parse(command);
 
-        this.api = new HttpClient(`${flags.host}:${flags.portAPI}/api/v2/`);
+        this.api = new HttpClient(`${flags.host}:${flags.portAPI}/api/`);
 
         await this.setupConfiguration();
         await this.setupConfigurationForCrypto();
@@ -202,8 +202,7 @@ export abstract class BaseCommand extends Command {
 
             this.constants = data.constants;
         } catch (error) {
-            logger.error(error.message);
-            process.exit(1);
+            this.error(error.message);
         }
     }
 
@@ -213,8 +212,7 @@ export abstract class BaseCommand extends Command {
 
             Managers.configManager.setConfig(dataCrypto);
         } catch (error) {
-            logger.error(error.message);
-            process.exit(1);
+            this.error(error.message);
         }
     }